+++ phpme/command/phpme_generate_constructor_command.py
@@ -28,7 +28,7 @@ class PhpmeGenerateConstructorCommand(sublime_plugin.TextCommand):
         if self.collect_progress == 0:
             if len(self.list_properties) > 0:
                 options = [
-                    ['Pick All', 'pick all properties'],
+                    ['Pick All', 'pick all properties ({} in total)'.format(len(self.list_properties))],
                     ['Pick Some', 'pick multiple properties one by one'],
                     ['Pick None', 'pick no property']
                 ]